Ulrika_Lookalike/requirements.txt
openai==1.12.0
httpx==0.27.0
flask==3.0.2
fpdf2==2.7.8
python-dotenv==1.0.0 